Add utilities for operating with version metadata
This commit introduces two new tools:

- lint-version-metadata.ts: Automatically check if both the
  `dlfx_var.shtml` and the `dltb2.shtml` are defined correctly.
- get-stable-version.ts: Get the latest stable version
  of Firefox and Thunderbird from API.

Both the tools can (in my opinion) largely improve the DX to
update the version of Firefox and Thunderbird,
while not being too automatic to preserve the
current version-updating workflow.

The both tools need only Deno runtime to run, and have limited
the permission they can retrieve. The detailed usage information
has embedded in the code, and you can refer to them when needed.

## To Be Discussed

- Should we document this tool in README.md?
  I'm not sure if README.md is suitable to place this.

Signed-off-by: pan93412 <pan93412@gmail.com>